[#34] OAuth 테스트 작성 (#40)
* refac: security 패키지 구조 변경
* fix: AccessDeniedException 메서드 이름 변경
* test: OAuth 테스트 작성
* refac: RefreshTokenRequest Validation
* refac: TokenResponse 레코드로 변경
* refac: OAuthService에 있는 UserRepository 의존성 제거
* feat: UserService 제작
* refac: OAuthController 패키지 구조 변경
* refac: provider 필드명 변경
* refac: TokenService createToken method signiture 변경
* feat: /token으로 끝나는 url는 jwtAuthentication filter 비적용
* refac: jwt 패키지 구조 변경
* refac: TokenResponse 변경에 의한 테스트 수정
* test: UserService 테스트 작성
* refac: token repository 패키지 변경
* refac: OAuthServiceTest 테스트 필드 Nested 클래스 안으로 변경
* refac: OAuthControllerTest에 사용되는 필드명과 위치 변경
