--- src/main/java/com/prgrms/amabnb/common/security/jwt/JwtAuthenticationToken.java
@@ -1,53 +1,0 @@
-package com.prgrms.amabnb.common.security.jwt;
-
-import java.util.Collection;
-
-import org.springframework.security.authentication.AbstractAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-public class JwtAuthenticationToken extends AbstractAuthenticationToken {
-
-    private final Object principal;
-    private String credentials;
-
-    public JwtAuthenticationToken(Object principal, String credentials) {
-        super(null);
-        super.setAuthenticated(false);
-        this.principal = principal;
-        this.credentials = credentials;
-    }
-
-    public JwtAuthenticationToken(
-        Object principal, String credentials, Collection<? extends GrantedAuthority> authorities
-    ) {
-        super(authorities);
-        super.setAuthenticated(true);
-        this.principal = principal;
-        this.credentials = credentials;
-    }
-
-    @Override
-    public Object getCredentials() {
-        return credentials;
-    }
-
-    @Override
-    public Object getPrincipal() {
-        return principal;
-    }
-
-    @Override
-    public void setAuthenticated(boolean authenticated) {
-        if (authenticated) {
-            throw new IllegalArgumentException("인증으로 변경할 수 없습니다.");
-        }
-        super.setAuthenticated(false);
-    }
-
-    @Override
-    public void eraseCredentials() {
-        super.eraseCredentials();
-        this.credentials = null;
-    }
-
-}
