--- src/main/java/com/prgrms/amabnb/common/security/jwt/exception/ExpiredTokenException.java
@@ -1,13 +1,0 @@
-package com.prgrms.amabnb.common.security.jwt.exception;
-
-import org.springframework.http.HttpStatus;
-
-public class ExpiredTokenException extends TokenException {
-
-    private static final String MESSAGE = "만료된 토큰입니다.";
-
-    public ExpiredTokenException() {
-        super(HttpStatus.UNAUTHORIZED, MESSAGE);
-    }
-
-}
