+++ src/main/java/com/prgrms/amabnb/oauth/dto/TokenResponse.java
@@ -1,19 +1,4 @@
 package com.prgrms.amabnb.oauth.dto;
 
-public class TokenResponse {
-    private final String accessToken;
-    private final String refreshToken;
-
-    public TokenResponse(String accessToken, String refreshToken) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-    }
-
-    public String getRefreshToken() {
-        return refreshToken;
-    }
-
-    public String getAccessToken() {
-        return accessToken;
-    }
+public record TokenResponse(String accessToken, String refreshToken) {
 }
