--- src/main/java/com/prgrms/amabnb/oauth/entity/TokenRepository.java
@@ -1,11 +1,0 @@
-package com.prgrms.amabnb.oauth.entity;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TokenRepository extends JpaRepository<Token, Long> {
-    
-    Optional<Token> findTokenByUserId(long userId);
-
-}
