+++ src/main/java/com/prgrms/amabnb/oauth/service/OAuthService.java
@@ -1,13 +1,11 @@
 package com.prgrms.amabnb.oauth.service;
 
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.prgrms.amabnb.oauth.dto.TokenResponse;
 import com.prgrms.amabnb.oauth.dto.UserProfile;
-import com.prgrms.amabnb.user.entity.User;
-import com.prgrms.amabnb.user.repository.UserRepository;
+import com.prgrms.amabnb.user.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,28 +15,12 @@ import lombok.RequiredArgsConstructor;
 public class OAuthService {
 
     private final TokenService tokenService;
-    private final UserRepository userRepository;
+    private final UserService userService;
 
     @Transactional
-    public TokenResponse register(String registerId, OAuth2User oauth) {
-        UserProfile userProfile = OAuthProvider.findByRegisterId(registerId)
-            .toUserProfile(oauth);
-        User user = userRepository.findByOauthId(String.valueOf(userProfile.getOauthId()))
-            .orElseGet(() -> {
-                    var newUser = createUser(registerId, oauth);
-                    return userRepository.save(newUser);
-                }
-            );
-
-        var userId = user.getId();
-        var role = user.getUserRole().getGrantedAuthority();
-        return tokenService.createToken(userId, role);
-    }
-
-    private User createUser(String registerId, OAuth2User oauth) {
-        UserProfile userProfile = OAuthProvider.findByRegisterId(registerId)
-            .toUserProfile(oauth);
-        return userProfile.toUser();
+    public TokenResponse register(UserProfile userProfile) {
+        var registeredUser = userService.register(userProfile);
+        return tokenService.createToken(registeredUser);
     }
 
 }
