+++ src/main/java/com/prgrms/amabnb/security/handler/JwtAuthenticationEntryPoint.java
@@ -1,0 +1,39 @@
+package com.prgrms.amabnb.security.handler;
+
+import java.io.IOException;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.stereotype.Component;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.prgrms.amabnb.common.exception.ErrorResponse;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+
+@Slf4j
+@Component
+@RequiredArgsConstructor
+public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
+
+    private static final String ERROR_MESSAGE = "로그인이 필요합니다.";
+    private static final String ERROR_LOG_MESSAGE = "[ERROR] {} : {}";
+
+    private final ObjectMapper objectMapper;
+
+    @Override
+    public void commence(HttpServletRequest request, HttpServletResponse response,
+        AuthenticationException authException) throws IOException {
+        log.info(ERROR_LOG_MESSAGE, authException.getClass().getSimpleName(), ERROR_MESSAGE);
+        response.setStatus(HttpStatus.UNAUTHORIZED.value());
+        response.setContentType("application/json");
+        response.setCharacterEncoding("UTF-8");
+        response.getWriter().write(objectMapper.writeValueAsString(new ErrorResponse(ERROR_MESSAGE)));
+    }
+
+}
