+++ src/main/java/com/prgrms/amabnb/user/dto/response/UserRegisterResponse.java
@@ -1,0 +1,4 @@
+package com.prgrms.amabnb.user.dto.response;
+
+public record UserRegisterResponse(Long id, String role) {
+}
