+++ src/main/java/com/prgrms/amabnb/user/repository/UserRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.prgrms.amabnb.user.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
+
     Optional<User> findByOauthId(String oauthId);
 
 }
